# olv-management/linux/app-api/openapi.yaml
openapi: 3.0.3
info:
  title: WireGuard App Client API
  description: |
    Public API cho iOS/macOS/Windows/Android app client kết nối WireGuard VPN.

    ## Authentication
    - Đăng nhập qua **Sign in with Apple** để nhận JWT token (30 ngày).
    - Tất cả endpoint trừ `/api/auth/*` và `/api/health` yêu cầu `Authorization: Bearer <token>`.

    ## Flow
    1. `GET /api/auth/me` → check token còn hợp lệ không (mở app lại)
    2. `POST /api/auth/apple` → nhận JWT (login lần đầu hoặc token hết hạn)
    3. `POST /api/devices` → đăng ký device (Secure Enclave public key)
    4. `GET /api/attest/challenge` → lấy challenge
    5. `POST /api/attest/verify` → gửi attestation từ `DCAppAttestService` (1 lần per device)
    6. `GET /api/servers` → xem danh sách server VPN
    7. `GET /api/attest/challenge` → lấy challenge mới
    8. `POST /api/connect` → tạo WireGuard peer (cần App Attest assertion headers)
    9. `GET /api/connect/status` → xem trạng thái kết nối

    ## App Attest (iOS only)
    `POST /api/connect` yêu cầu **Apple App Attest assertion** cho **iOS devices**.
    macOS không hỗ trợ `DCAppAttestService` nên được skip.
    - iOS device phải được attest trước qua `/api/attest/verify` (1 lần duy nhất)
    - Mỗi lần gọi `/api/connect` từ iOS cần: challenge mới + assertion headers
    - macOS gọi `/api/connect` bình thường không cần attest headers
  version: 1.0.0
  contact:
    name: WireGuard Management
servers:
  - url: https://localhost
...
  - url: https://api.example.com
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from `/api/auth/apple`
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Something went wrong
...
    ConnectionConfig:
      type: object
      properties:
        peerId:
          type: integer
          description: Peer metadata ID
          example: 42
        assignedIp:
          type: string
          description: IP address assigned to this peer
          example: "10.0.0.5/32"
        serverPublicKey:
          type: string
          format: byte
          description: WireGuard server public key (Curve25519, base64)
          example: "xTIBA5rboUvnH4htodjb6e697QjLERt1NAB4mZqp8Dg="
        serverEndpoint:
          type: string
          description: "Server hostname:port"
          example: "vpn-sg.example.com:51820"
        allowedIPs:
          type: string
          description: |
            Comma-separated CIDR ranges that should be routed through the VPN tunnel.
            Client MUST use this value for the peer's `AllowedIPs` in WireGuard config.

            - Default (no admin config): subnet gateway `/32`, e.g. `"10.0.0.1/32"`
            - Admin configured: custom routes, e.g. `"10.0.0.0/8, 192.168.1.0/24"`
          example: "10.0.0.1/32"
        dns:
          type: string
          description: DNS server
          example: "1.1.1.1"
        persistentKeepalive:
          type: integer
          description: Keepalive interval in seconds
          example: 25
        serverName:
          type: string
          description: Human-readable server name
          example: "SG-01"
    Connection:
      type: object
      properties:
        peerId:
          type: integer
        serverName:
          type: string
          example: "SG-01"
        interfaceName:
          type: string
          example: "wg0"
        publicKey:
          type: string
          format: byte
          description: WireGuard public key of this peer
        deviceName:
          type: string
          nullable: true
          example: "iPhone 15 Pro"
        deviceOs:
          type: string
          nullable: true
          enum: [macos, ios, windows, android]
        alias:
          type: string
          example: "user@icloud.com/iPhone 15 Pro"
        createdAt:
          type: string
          format: date-time
        isExpired:
          type: boolean
        connected:
          type: boolean
          description: true if latest handshake < 180 seconds ago
        latestHandshake:
          type: number
          nullable: true
          description: Unix timestamp of latest WireGuard handshake
...
paths:
...
  /api/connect:
    post:
      tags: [Connect]
      summary: Create WireGuard peer (connect VPN)
      operationId: createConnection
      security:
        - BearerAuth: []
      description: |
        Tạo WireGuard peer cho device. Mỗi device chỉ có 1 peer — peer cũ tự động bị xoá.

        **Yêu cầu App Attest assertion cho iOS** — device iOS phải đã attest qua `/api/attest/verify`.
        macOS devices không cần App Attest headers (tự động skip).

        **Flow:**
        1. `GET /api/attest/challenge` → nhận challenge mới
        2. App tạo WireGuard keypair (Curve25519)
        3. App dùng Secure Enclave private key ký `wgPublicKey` → `signature`
        4. App gọi `generateAssertion(keyId, SHA256(JSON.stringify(body)))` → assertion
        5. Gửi request với assertion headers + challenge trong body
        6. Server verify App Attest assertion → verify SE signature → tạo peer

        **Server selection:**
        - Nếu gửi `serverId` → dùng server đó
        - Nếu không → auto-select server có ít peer nhất

        **iOS/macOS client:**
        ```swift
        let challenge = // GET /api/attest/challenge
        let body: [String: Any] = [
          "deviceId": deviceId,
          "wgPublicKey": wgPubKeyBase64,
          "signature": signatureBase64,
          "challenge": challenge
        ]
        let bodyData = try JSONSerialization.data(withJSONObject: body)
        let clientDataHash = SHA256.hash(data: bodyData)
        let assertion = try await DCAppAttestService.shared.generateAssertion(keyId, clientDataHash: Data(clientDataHash))
        // POST /api/connect with headers + body
        ```
      parameters:
        - name: X-App-Attest-KeyId
          in: header
          required: false
          schema:
            type: string
          description: Key ID from App Attest attestation (required for iOS, ignored for macOS)
        - name: X-App-Attest-Assertion
          in: header
          required: false
          schema:
            type: string
            format: byte
          description: Base64-encoded assertion from `generateAssertion()` (required for iOS, ignored for macOS)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - deviceId
                - wgPublicKey
                - signature
              properties:
                deviceId:
                  type: string
                  format: uuid
                  description: Device ID from `/api/devices`
                wgPublicKey:
                  type: string
                  format: byte
                  description: Base64-encoded WireGuard Curve25519 public key
                signature:
                  type: string
                  format: byte
                  description: Base64-encoded ECDSA P-256 signature of wgPublicKey UTF-8 bytes (signed by Secure Enclave)
                presharedKey:
                  type: string
                  format: byte
                  nullable: true
                  description: Base64-encoded WireGuard preshared key (optional)
                serverId:
                  type: integer
                  nullable: true
                  description: Target server ID (optional, auto-select if omitted)
                challenge:
                  type: string
                  description: Challenge from `GET /api/attest/challenge` (required for iOS, optional for macOS)
      responses:
        "201":
          description: Peer created, return connection config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionConfig"
        "400":
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "deviceId, wgPublicKey, and signature are required"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: App Attest failed, device disabled, or invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                attest_missing:
                  value:
                    error: "App Attest required: X-App-Attest-KeyId, X-App-Attest-Assertion headers and challenge in body"
                attest_failed:
                  value:
                    error: "App Attest assertion failed"
                not_attested:
                  value:
                    error: "Device not attested. Call /api/attest/verify first."
                disabled:
                  value:
                    error: "Device is not enabled"
                bad_signature:
                  value:
                    error: "Invalid signature"
        "404":
          description: Device or server not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                device:
                  value:
                    error: "Device not found"
                server:
                  value:
                    error: "Server not found"
        "409":
          description: No subnet or no available IPs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                no_subnet:
                  value:
                    error: "No subnet configured on this server"
                no_ip:
                  value:
                    error: "No available IPs in subnet"
        "503":
          description: No servers available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "No servers available"
  /api/connect/status:
    get:
      tags: [Connect]
      summary: Get active VPN connections
      operationId: getConnectionStatus
      security:
        - BearerAuth: []
      description: |
        Trả về danh sách kết nối VPN hiện tại của user, kèm trạng thái live.

        `connected: true` nếu WireGuard handshake gần nhất < 180 giây.
      responses:
        "200":
          description: Connection list with live status
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Connection"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"